Move grad test to tracing.test.ts
