+++ test/tracing.test.ts
@@ -428,6 +428,22 @@ suite("jax.grad()", () => {
     ]);
   });
 
+  test("can differentiate through broadcast of a size-1 axis", () => {
+    // The transpose of a broadcast must sum cotangents over expanded axes of
+    // size 1, not just over newly inserted axes.
+    const f = (x: np.Array) =>
+      np.sum(
+        np.broadcastTo(x.reshape([2, 1]), [2, 3]).mul(
+          np.array([
+            [1, 2, 3],
+            [4, 5, 6],
+          ]),
+        ),
+      );
+    const dx = grad(f)(np.array([1.0, 2.0]));
+    expect(dx).toBeAllclose([6, 15]);
+  });
+
   test("hasAux returns aux and computes correct gradient", () => {
     const f = (x: np.Array): [np.Array, np.Array] => {
       const loss = x.ref.sum();
